stopping work for now, changes in design
