+++ src/js/components/covid19/amountsVisualization/AmountsVisualization.jsx
@@ -7,7 +7,7 @@ import React, { useState, useEffect, useRef, useLayoutEffect } from 'react';
 import PropTypes from 'prop-types';
 import { upperFirst, throttle } from 'lodash';
 import { scaleLinear } from 'd3-scale';
-import TotalBudgetaryResources from 'components/covid19/amountsVisualization/amounts/TotalBudgetaryResources';
+import TotalBudgetaryResources from 'components/covid19/amountsVisualization/amounts/totalBudgetaryResources/TotalBudgetaryResources';
 import DateNote from 'components/covid19/DateNote';
 import ResultsTableLoadingMessage from 'components/search/table/ResultsTableLoadingMessage';
 import { TooltipWrapper } from 'data-transparency-ui';
@@ -214,73 +214,6 @@ const AmountsVisualization = ({
             if (!isNaN(scale(amount))) setOutlayLineData(data);
         }
     }, [scale, overviewData]);
-    // obligationLineDataOne
-    useEffect(() => {
-        if (scale) {
-            const {
-                offset,
-                lineColor
-            } = rectangleMapping._totalObligations;
-            const { left, right } = amountsPadding;
-            const amount = Math.abs(overviewData._totalObligations);
-            const x = left + offset.left;
-            const rectWidth = scale(amount) - (right + (offset.right || 0));
-            const data = {
-                lineColor,
-                x1: (x + rectWidth) - (lineStrokeWidth / 2),
-                x2: (x + rectWidth) - (lineStrokeWidth / 2),
-                y1: startOfChartY + (rectangleHeight / 2),
-                y2: startOfChartY + rectangleHeight + heightOfRemainingBalanceLines + (spacingBetweenLineAndText / 2)
-            };
-            if (!isNaN(scale(amount))) setObligationLineDataOne(data);
-        }
-    }, [scale, overviewData]);
-    // obligationLineDataTwo
-    useEffect(() => {
-        if (scale) {
-            const amount = Math.abs(overviewData._totalObligations);
-            const data = {
-                lineColor: rectangleMapping._totalObligations.lineColor,
-                x1: obligationLineDataOne?.x1 || 0,
-                x2: obligationLineDataOne?.x1 || 0,
-                y1: obligationLineDataOne?.y2 || 0,
-                y2: (obligationLineDataOne?.y2 || 0) + (obligationQuestionData?.height || 0) + (rectangleMapping._remainingBalance.text.offset.y / 2) + 3
-            };
-            if (!isNaN(scale(amount))) setObligationLineDataTwo(data);
-        }
-    }, [scale, obligationLineDataOne]);
-    // obligationLineDataThree
-    useEffect(() => {
-        if (scale) {
-            const amount = Math.abs(overviewData._totalObligations);
-            const data = {
-                lineColor: rectangleMapping._totalObligations.lineColor,
-                x1: obligationLineDataTwo?.x1 || 0,
-                x2: obligationLineDataTwo?.x1 || 0,
-                y1: obligationLineDataTwo?.y2 || 0,
-                y2: ((obligationLineDataTwo?.y2 || 0) + (obligationValueData?.height || 0)) - 3
-            };
-            if (!isNaN(scale(amount))) setObligationLineDataThree(data);
-        }
-    }, [scale, obligationLineDataTwo]);
-    // obligationLineDataFour
-    useEffect(() => {
-        if (scale) {
-            const {
-                lineLength,
-                lineColor
-            } = rectangleMapping._totalObligations;
-            const amount = Math.abs(overviewData._totalObligations);
-            const data = {
-                lineColor,
-                x1: obligationLineDataThree?.x1 || 0,
-                x2: obligationLineDataThree?.x1 || 0,
-                y1: obligationLineDataThree?.y2 || 0,
-                y2: startOfChartY + rectangleHeight + lineLength
-            };
-            if (!isNaN(scale(amount))) setObligationLineDataFour(data);
-        }
-    }, [scale, obligationLineDataThree]);
     // remainingBalanceQuestionData
     useLayoutEffect(() => {
         const { text: textInfo } = rectangleMapping._remainingBalance;
@@ -365,71 +298,6 @@ const AmountsVisualization = ({
             });
         }
     }, [outlayValueData]);
-    // obligationQuestionData
-    useLayoutEffect(() => {
-        const { text: textInfo } = rectangleMapping._totalObligations;
-        const questionRef = _obligationQuestion.current?.getBoundingClientRect();
-        setObligationQuestionData({
-            y: (obligationLineDataFour?.y2 || 0) - spacingBetweenLineAndText,
-            x: (obligationLineDataFour?.x1 || 0) - ((questionRef?.width || 0) + spacingBetweenLineAndText),
-            height: questionRef?.height || 0,
-            text: textInfo.question,
-            className: `amounts-text__question ${!questionRef ? 'white' : ''}`,
-            left: true
-        });
-    }, [obligationLineDataFour]);
-    // obligationValueData
-    useLayoutEffect(() => {
-        const ref = _obligationValue.current?.getBoundingClientRect();
-        const amount = Math.abs(overviewData._totalObligations);
-        const units = calculateUnits([amount]);
-        const moneyLabel = `${formatMoneyWithPrecision(amount / units.unit, 1)} ${upperFirst(units.longLabel)}`;
-        if (obligationQuestionData?.left) {
-            setObligationValueData({
-                y: (obligationLineDataFour?.y2 || 0) - (obligationQuestionData?.height || 0) - spacingBetweenLineAndText,
-                x: (obligationLineDataFour?.x1 || 0) - ((ref?.width || 0) + spacingBetweenLineAndText),
-                height: ref?.height || 0,
-                theWidth: ref?.width || 0,
-                text: moneyLabel,
-                className: `amounts-text__value ${!ref ? 'white' : ''}`
-            });
-        }
-        else {
-            setObligationValueData({
-                y: (obligationLineDataFour?.y2 || 0) - (obligationQuestionData?.height || 0) - spacingBetweenLineAndText,
-                x: (obligationLineDataFour?.x1 || 0) + (obligationLabelData?.theWidth || 0) + spacingBetweenLineAndText + labelTextAdjustment.x,
-                height: ref?.height || 0,
-                theWidth: ref?.width || 0,
-                text: moneyLabel,
-                className: `amounts-text__value ${!ref ? 'white' : ''}`
-            });
-        }
-    }, [obligationLineDataFour, zeroPercentData]);
-    // obligationLabelData
-    useLayoutEffect(() => {
-        const ref = _obligationLabel.current?.getBoundingClientRect();
-        const { text: textInfo } = rectangleMapping._totalObligations;
-        if (obligationQuestionData?.left) {
-            setObligationLabelData({
-                y: (obligationLineDataFour?.y2 || 0) - (obligationQuestionData?.height || 0) - spacingBetweenLineAndText,
-                x: (obligationLineDataFour?.x1 || 0) - ((ref?.width || 0) + (obligationValueData?.theWidth || 0) + spacingBetweenLineAndText + labelTextAdjustment.x),
-                height: ref?.height || 0,
-                text: textInfo.label,
-                theWidth: ref?.width || 0,
-                className: `amounts-text__label ${!ref ? 'white' : ''}`
-            });
-        }
-        else {
-            setObligationLabelData({
-                y: (obligationLineDataFour?.y2 || 0) - spacingBetweenLineAndText,
-                x: (obligationLineDataFour?.x1 || 0) + spacingBetweenLineAndText,
-                height: ref?.height || 0,
-                text: textInfo.label,
-                theWidth: ref?.width || 0,
-                className: `amounts-text__label ${!ref ? 'white' : ''}`
-            });
-        }
-    }, [obligationValueData]);
     // 0%
     useLayoutEffect(() => {
         const ref = zerPercentRef.current?.getBoundingClientRect();
@@ -498,24 +366,6 @@ const AmountsVisualization = ({
             setRemainingBalanceHorizontalLineData(data);
         }
     }, [drawRemainingBalanceItems]);
-    // obligationLineDataTwoOverlap
-    useEffect(() => {
-        if (((obligationLineDataTwo?.x1 || 0) + (lineStrokeWidth / 2)) >= (remainingBalanceLabelData?.x || 0)) {
-            setObligationLineDataTwoOverlap(true);
-        }
-        else {
-            setObligationLineDataTwoOverlap(false);
-        }
-    }, [obligationLineDataTwo, remainingBalanceLabelData, remainingBalanceValueData]);
-    // obligationLineDataThreeOverlap
-    useEffect(() => {
-        if (((obligationLineDataThree?.x1 || 0) + (lineStrokeWidth / 2)) >= (remainingBalanceQuestionData?.x || 0)) {
-            setObligationLineDataThreeOverlap(true);
-        }
-        else {
-            setObligationLineDataThreeOverlap(false);
-        }
-    }, [obligationLineDataThree, remainingBalanceQuestionData]);
     // draw remaining balance text
     useEffect(() => {
         if (overviewData) {
@@ -601,87 +451,7 @@ const AmountsVisualization = ({
                         overviewData={overviewData}
                         scale={scale} />
                     {
-                        obligationLineDataOne &&
-                        <g
-                            tabIndex="0"
-                            aria-label="A line linking a rectangle to text"
-                            data-id="_totalObligations"
-                            onFocus={displayTooltip}
-                            onBlur={hideTooltip}>
-                            <desc>A line linking a rectangle to text</desc>
-                            <line
-                                className="line__obligation"
-                                x1={obligationLineDataOne.x1}
-                                data-id="_totalObligations"
-                                x2={obligationLineDataOne.x2}
-                                y1={obligationLineDataOne.y1}
-                                y2={obligationLineDataOne.y2}
-                                strokeWidth={lineStrokeWidth}
-                                onMouseMove={displayTooltip}
-                                onMouseLeave={hideTooltip} />
-                        </g>
-                    }
-                    {
-                        obligationLineDataTwo &&
-                        <g
-                            tabIndex="0"
-                            aria-label="A line 2 linking a rectangle to text"
-                            data-id="_totalObligations"
-                            onFocus={displayTooltip}
-                            onBlur={hideTooltip}>
-                            <desc>A line linking a rectangle to text</desc>
-                            <line
-                                x1={obligationLineDataTwo.x1}
-                                data-id="_totalObligations"
-                                x2={obligationLineDataTwo.x2}
-                                y1={obligationLineDataTwo.y1}
-                                y2={obligationLineDataTwo.y2}
-                                className={(drawRemainingBalanceText && obligationLineDataTwoOverlap) ? 'line__opacity' : 'line__obligation'}
-                                onMouseMove={displayTooltip}
-                                onMouseLeave={hideTooltip} />
-                        </g>
-                    }
-                    {
-                        obligationLineDataThree &&
-                        <g
-                            tabIndex="0"
-                            aria-label="A line 3 linking a rectangle to text"
-                            data-id="_totalObligations"
-                            onFocus={displayTooltip}
-                            onBlur={hideTooltip}>
-                            <desc>A line linking a rectangle to text</desc>
-                            <line
-                                x1={obligationLineDataThree.x1}
-                                data-id="_totalObligations"
-                                x2={obligationLineDataThree.x2}
-                                y1={obligationLineDataThree.y1}
-                                y2={obligationLineDataThree.y2}
-                                className={(drawRemainingBalanceText && obligationLineDataThreeOverlap) ? 'line__opacity' : 'line__obligation'}
-                                onMouseMove={displayTooltip}
-                                onMouseLeave={hideTooltip} />
-                        </g>
-                    }
-                    {
-                        obligationLineDataFour &&
-                        <g
-                            tabIndex="0"
-                            aria-label="A line 4 linking a rectangle to text"
-                            data-id="_totalObligations"
-                            onFocus={displayTooltip}
-                            onBlur={hideTooltip}>
-                            <desc>A line linking a rectangle to text</desc>
-                            <line
-                                className="line__obligation"
-                                data-id="_totalObligations"
-                                x1={obligationLineDataFour.x1}
-                                x2={obligationLineDataFour.x2}
-                                y1={obligationLineDataFour.y1}
-                                y2={obligationLineDataFour.y2}
-                                strokeWidth={lineStrokeWidth}
-                                onMouseMove={displayTooltip}
-                                onMouseLeave={hideTooltip} />
-                        </g>
-                    }
+
                     {
                         obligationRectangleData &&
                         <g
@@ -897,69 +667,6 @@ const AmountsVisualization = ({
                                 onMouseMove={displayTooltip}
                                 onMouseLeave={hideTooltip}>
                                 {outlayLabelData.text}
-                            </text>
-                        </g>
-                    }
-                    {
-                        obligationQuestionData &&
-                        <g
-                            tabIndex="0"
-                            aria-label={obligationQuestionData.text}
-                            onFocus={displayTooltip}
-                            data-id="_totalObligations"
-                            onBlur={hideTooltip}>
-                            <desc>{obligationQuestionData.text}</desc>
-                            <text
-                                ref={_obligationQuestion}
-                                data-id="_totalObligations"
-                                className={obligationQuestionData.className}
-                                x={obligationQuestionData.x}
-                                y={obligationQuestionData.y}
-                                onMouseMove={displayTooltip}
-                                onMouseLeave={hideTooltip}>
-                                {obligationQuestionData.text}
-                            </text>
-                        </g>
-                    }
-                    {
-                        obligationValueData &&
-                        <g
-                            tabIndex="0"
-                            aria-label={obligationValueData.text}
-                            data-id="_totalObligations"
-                            onFocus={displayTooltip}
-                            onBlur={hideTooltip}>
-                            <desc>{obligationValueData.text}</desc>
-                            <text
-                                ref={_obligationValue}
-                                data-id="_totalObligations"
-                                className={obligationValueData.className}
-                                x={obligationValueData.x}
-                                y={obligationValueData.y}
-                                onMouseMove={displayTooltip}
-                                onMouseLeave={hideTooltip}>
-                                {obligationValueData.text}
-                            </text>
-                        </g>
-                    }
-                    {
-                        obligationLabelData &&
-                        <g
-                            tabIndex="0"
-                            aria-label={obligationLabelData.text}
-                            data-id="_totalObligations"
-                            onFocus={displayTooltip}
-                            onBlur={hideTooltip}>
-                            <desc>{obligationLabelData.text}</desc>
-                            <text
-                                ref={_obligationLabel}
-                                data-id="_totalObligations"
-                                className={obligationLabelData.className}
-                                x={obligationLabelData.x}
-                                y={obligationLabelData.y}
-                                onMouseMove={displayTooltip}
-                                onMouseLeave={hideTooltip}>
-                                {obligationLabelData.text}
                             </text>
                         </g>
                     }
